feat: update the site to the newest Jekyll
I created a fresh Jekyll instance and migrated the setup from there
here. The biggest difference is that Jekyll now uses plugins for themes,
i.e., the minima theme files are not tracked by my repo.

I needed to modify the CSS for figures (_sass files) and add personal
icons to the socials.